64-bit, hex bitwise ops, unary ops, 's' sign, better error checking.
Tested with CentOS 5, CentOS, 6.5, and CentOS 7.  Testing required
several source code changes in order to get clean builds and in order to
get the tests to pass.

Added the first few hex bitwise operators (AND, OR, XOR, NOT).

Support for unary operators (i.e. NOT).

Added 's', which allows the user to enter negative numbers ('s' stands
for sign.  Using 's' instead of '-' prevents confusion between negative
numbers and the subtraction symbol).

Converted a lot of members to return a bool, and then check the return
codes in order to watch for internal failures.