For ints, operand_get_val() returns float and int.  New operand_to_str().
operand_get_val() used to return an int for int operands and a float for
float operands.  Now, if an operand is an int, it returns a float and an
int.  This allows the caller to use to i_val or f_val.

Added new function operand_to_str().  Now the operand class owns the
right to generate the output of the operands.