Major cleanup of TODO list.
I removed all of the tasks that have been completed.  I also cleaned up
some tasks that are not completed.  Some of them are no longer relevent,
so they were removed.  Added a few that were missing.  The TODO list now
does a better job of listing the remaining work.