First pass at converting fp_exp.[ch] to use BCD.
Previously fp_exp used "double" (IEEE 754 floating point), which isn't
completely accurate for a calculator.  So the entire calculator has been
switched from IEEE 754 to BCD.

Note that fp_exp still has problems that need to be fixed.  But it
compiles and runs and does a lot of simple exponentiation correctly.
There are unit tests in place that fail because of the known bugs.  So
the module won't be declared functionally complete until all of the
functional tests pass.