Fixed rounding in bcd_to_str().
We now round up if necessary.  Now that we have double precision data we
can see what the next digit (after the final one we display) is supposed
to be, so we're able to round up appropriately.

Note that we don't modify the contents of "this".  We create a copy of
this, and then we round it.  This allows us to continue to use the full
value of this in subsequent calculations.