Added an elapsed_time.c program.  Simplified the prime_numbers test.
The elapsed_time.c program simply calculates the elapsed time for a
task.

The prime_numbers test uses a simpler mechanism for calculating the
elapsed time.  It is displayed as the number of CPU cycles used to
perform a test.  But the overall prime_numbers test is still a simple
brute force test.