Intermediate commit.  Switched to Makefile and big_number_sim.c.
The Makefile is just a better way to build.

big_number_sim.c is a simulation of the real big_number class.  It uses
uint64_t variables, which allows us to use regular C math operations.
The real big_number class won't be able to use the C math.  So
big_number_sim provides an easy way to do test/debug of the crypto_algs
app.