Added shm_fork_rwlock to the shared memory examples.
This program uses a read-write lock to control activities between
processes.  The parent will occasionally grab the write lock and modify
some data in a shared memory region.  The child processes grab a read
lock and read the data.  The program moves slowly so that you can see the
behavior of the read-write lock.